+++ packages/layers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mapwise/layers
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @mapwise/core@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes